adding STM campus to cvs
Note that the current floor areas in CSV and geo-json for STM are
footprint areas, I think based on footprint drawing, not floor area.
Both of those files should be updated with proper values. Also the
values in this appear to be IP values while the prototype footprint is
SI.